requirements.txt
langchain
langgraph
pandas
python-pptx
openai
streamlit
langchain_community
python-dotenv